Ajustes
Correções no projeto

+++ Scripts/tarefas.js
@@ -3,11 +3,32 @@
 var Tarefa = function (descricao, concluida) {
     this.descricao = descricao;
     this.concluida = concluida;
+    this.usuario = pegarUsuarioAutorizado();
 };
 
 function buscarTarefas() {
     var tarefas = localStorage.getItem(chaveTarefas);
+
     return tarefas != null ? JSON.parse(tarefas) : new Array();
+}
+
+function buscarTarefasPorUsuario() {
+    var tarefas = localStorage.getItem(chaveTarefas);
+
+    if (tarefas == null)
+        return new Array();
+
+    return $.grep(JSON.parse(tarefas), function (item) {
+        return item.usuario.email === pegarUsuarioAutorizado().email;
+    });    
+}
+
+function buscarTarefasPendentesDoUsuario() {
+    var tarefasDoUsuario = buscarTarefasPorUsuario();
+
+    return $.grep(tarefasDoUsuario, function(item) {
+        return item.concluida == false;
+    });
 }
 
 function cadastrarTarefa(descricao, concluida) {
@@ -15,9 +36,16 @@ function cadastrarTarefa(descricao, concluida) {
 
     var tarefas = buscarTarefas();
 
+    if (tarefaJaExiste(tarefa, tarefas)) {
+        alert("Já existe uma tarefa com a mesma descrição.");
+        return false;
+    }
+
     tarefas.push(tarefa);
 
     localStorage.setItem(chaveTarefas, JSON.stringify(tarefas));
+
+    return true;
 }
 
 function atualizarTarefa(tarefa) {
@@ -26,7 +54,7 @@ function atualizarTarefa(tarefa) {
     var indiceDaTarefa = -1;
 
     $.each(tarefas, function(indice, item) {
-        if (item.descricao.trim() === tarefa.descricao.trim()) {
+        if (item.descricao.trim() === tarefa.descricao.trim() && item.usuario.email === pegarUsuarioAutorizado().email) {
             indiceDaTarefa = indice;
             return;
         }
@@ -38,4 +66,12 @@ function atualizarTarefa(tarefa) {
     tarefas[indiceDaTarefa].concluida = tarefa.concluida;
 
     localStorage.setItem(chaveTarefas, JSON.stringify(tarefas));
+}
+
+function tarefaJaExiste(tarefa, tarefas) {
+    var tarefaEncontrada = $.grep(tarefas, function (item) {
+        return item.descricao.trim() === tarefa.descricao.trim() && item.usuario.email === pegarUsuarioAutorizado().email;
+    });
+
+    return tarefaEncontrada.length > 0;
 }
